excel-op 2/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pandas==2.2.3
openpyxl==3.1.5

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3

chromadb==0.5.11
sentence-transformers>=2.2.0

crewai
langchain
langchain-openai
openai

python-dotenv==1.0.1
pydantic-settings==2.6.1

pytest==8.3.3
httpx==0.27.2
ruff==0.7.3
black==24.10.0
